requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
cloudinary==1.36.0
Django==4.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.5
whitenoise==6.5.0
